+++ lib/int_part1.f
@@ -43,7 +43,7 @@
          r2 = xx*xx + yy*yy
          arg = r2*s2inv2
          if (arg < gauss_cut) then
-            i = arg/gdelta + 1
+            i = int(arg/gdelta + 1)
             c = gauss(i)*(1 + i*gdelta - arg) ! includes first error term
          else
             c = 0.
